TempHum Sending correct Packets!!!!!!!
First sensor working fully so it's only up from here :D

+++ Final/CAN.cydsn/codegentemp/ADC_EOC_Int.c
@@ -1,6 +1,6 @@
 /*******************************************************************************
 * File Name: ADC_EOC_Int.c  
-* Version 1.70
+* Version 1.71
 *
 *  Description:
 *   API for controlling the state of an interrupt.
@@ -252,7 +252,7 @@ void ADC_EOC_Int_SetPriority(uint8 priority)
     uint32 priorityOffset = ((ADC_EOC_Int__INTC_NUMBER % 4u) * 8u) + 6u;
     
 	interruptState = CyEnterCriticalSection();
-    *ADC_EOC_Int_INTC_PRIOR = (*ADC_EOC_Int_INTC_PRIOR & (uint32)(~ADC_EOC_Int__INTC_PRIOR_MASK)) |
+    *ADC_EOC_Int_INTC_PRIOR = (*ADC_EOC_Int_INTC_PRIOR & (uint32)(~(uint32)ADC_EOC_Int__INTC_PRIOR_MASK)) |
                                     ((uint32)priority << priorityOffset);
 	CyExitCriticalSection(interruptState);
 }
